LayoutTests/fast/events/onunload-expected.txt

main frame - has 1 onunload handler(s)
CONSOLE MESSAGE: line 16: Use of window.alert is not allowed while unloading a page.
You should have seen an unload alert appear.
